Moving TokenCard
