+++ src/components/TokenCard/TokenCard.tsx
@@ -3,8 +3,10 @@ import { View } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 import { Text, Icon, Token, Input } from '@components';
 import { TokenType } from '@styles';
+import { useTheme } from '@hooks';
 import { TokenCardProps } from './TokenCard.types';
 import { useTokenCard } from './TokenCard.hooks';
+import { makeStyles } from './TokenCard.styles';
 
 const TokenCard: React.FC<TokenCardProps> = ({
 	token,
@@ -15,7 +17,10 @@ const TokenCard: React.FC<TokenCardProps> = ({
 	updateQuotes,
 	conversionAmount = ''
 }) => {
-	const { amount, onChangeText, onMaxPress, isMaxEnabled, invalidAmount, styles } = useTokenCard({
+	const { colors } = useTheme();
+	const styles = makeStyles(colors);
+
+	const { amount, onChangeText, onMaxPress, isMaxEnabled, invalidAmount } = useTokenCard({
 		balance,
 		updateQuotes: updateQuotes!,
 		token: token!,
